Move UpdateClientConfigResult to CorePayments.
