+++ CorePayments/src/main/java/com/paypal/android/corepayments/UpdateClientConfigParams.kt
@@ -1,0 +1,12 @@
+package com.paypal.android.corepayments
+
+import androidx.annotation.RestrictTo
+
+/**
+ * @suppress
+ */
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+data class UpdateClientConfigParams(
+    val orderId: String,
+    val fundingSource: String
+)
