+++ CorePayments/src/main/res/raw/graphql_query_update_client_config.graphql
